Add files via upload
This is a chat application by which two localhosts can communicate with each other using three laptops and connected via common WIFI or maybe via common mobile hotspot. In that one PC will work as a server and rest other two will work as clients.
The thing that matters the most is that all three should be connected to the same IP and same port no.